+++ packages/form/src/components/FieldSet/demos/components-other.tsx
@@ -52,6 +52,9 @@ const Demo = () => (
       />
       <ProFormSelect
         width="md"
+        fieldProps={{
+          labelInValue: true,
+        }}
         request={async () => [
           { label: '全部', value: 'all' },
           { label: '未解决', value: 'open' },
@@ -65,6 +68,9 @@ const Demo = () => (
         <ProFormSelect.SearchSelect
           name="userQuery"
           label="查询选择器"
+          fieldProps={{
+            labelInValue: true,
+          }}
           request={async () => [
             { label: '全部', value: 'all' },
             { label: '未解决', value: 'open' },
@@ -116,7 +122,7 @@ const Demo = () => (
         placeholder="Please select favorite colors"
         rules={[{ required: true, message: 'Please select your favorite colors!', type: 'array' }]}
       />
-      <ProFormDigit label="InputNumber" name="input-number" min={1} max={10} />
+      <ProFormDigit label="InputNumber" name="input-number" width="sm" min={1} max={10} />
       <ProFormSwitch name="switch" label="Switch" />
       <ProFormSlider
         name="slider"