+++ packages/utils/src/typing.ts
@@ -48,16 +48,16 @@ export type ProFieldValueType =
   | 'image'
   | 'jsonCode';
 
-export type ProFieldRequestData<U = any> = (
-  params: U,
-  props: any,
-) => Promise<
-  {
-    label?: React.ReactNode;
-    value?: React.ReactText;
-    [key: string]: any;
-  }[]
->;
+export type RequestOptionsType = {
+  label?: React.ReactNode;
+  value?: React.ReactText;
+  /** 渲染的节点类型 */
+  optionType?: 'optGroup' | 'option';
+  children?: Omit<RequestOptionsType, 'children' | 'optionType'>[];
+  [key: string]: any;
+};
+
+export type ProFieldRequestData<U = any> = (params: U, props: any) => Promise<RequestOptionsType[]>;
 
 export type ProFieldValueEnumType = ProSchemaValueEnumMap | ProSchemaValueEnumObj;
 
@@ -72,7 +72,6 @@ export type ProFieldValueObjectType = {
   precision?: number;
   moneySymbol?: string;
   request?: ProFieldRequestData;
-
   /** Image */
   width?: number;
 };